requirements.txt
azure-functions
requests
tenacity
azure-eventhub